feat(webpack-builder): re-organizing config default value types
