+++ packages/builder/webpack-builder/src/config/defaults.ts
@@ -1,4 +1,4 @@
-import { BuilderConfig } from '../types';
+import { BuilderConfig, Config } from '../types';
 import { mergeBuilderConfig } from '../shared/utils';
 import {
   ROOT_DIST_DIR,
@@ -115,33 +115,10 @@ const defaultConfig = defineConfigPreserveDetails({
 
 export type DefaultConfig = typeof defaultConfig;
 
-type KeysMatching<T extends object, V> = {
-  [K in keyof T]-?: T[K] extends V ? K : never;
-}[keyof T];
-
-type ObjectProps<T extends Record<any, any>> = Pick<
-  T,
-  KeysMatching<T, Record<any, any>>
->;
-
-export type DeepFillObjectBy<T extends object, P extends T> = {
-  [K in keyof T & keyof ObjectProps<P>]-?: Exclude<
-    Exclude<T[K], undefined> extends object
-      ? unknown &
-          DeepFillObjectBy<Exclude<T[K], undefined>, Exclude<P[K], undefined>>
-      : T[K],
-    undefined
-  >;
-} & {
-  [K in Exclude<keyof T, keyof ObjectProps<P>>]: T[K] extends Record<any, any>
-    ? unknown & DeepFillObjectBy<T[K], P[K]>
-    : T[K];
-};
-
-export type Config = DeepFillObjectBy<BuilderConfig, DefaultConfig>;
-
-export const cloneDefaultConfig = (): BuilderConfig =>
-  _.cloneDeep(defaultConfig);
+export const cloneDefaultConfig = (): Config =>
+  // typescript unable to match the type of defaultConfig
+  _.cloneDeep(defaultConfig) as unknown as Config;
 
 export const withDefaultConfig = (config: BuilderConfig) =>
-  mergeBuilderConfig(cloneDefaultConfig(), config);
+  // typescript unable to match the type of defaultConfig
+  mergeBuilderConfig(cloneDefaultConfig(), config as any);
