+++ packages/builder/webpack-builder/src/types/config/index.ts
@@ -6,6 +6,8 @@ import type { OutputConfig } from './output';
 import type { SecurityConfig } from './security';
 import type { PerformanceConfig } from './performance';
 import type { ExperimentsConfig } from './experiments';
+import type { DefaultConfig } from '../../config/defaults';
+import type { DeepFillObjectBy } from '../utils';
 
 export interface BuilderConfig {
   dev?: DevConfig;
@@ -17,6 +19,9 @@ export interface BuilderConfig {
   performance?: PerformanceConfig;
   experiments?: ExperimentsConfig;
 }
+
+export type Config = DeepFillObjectBy<BuilderConfig, DefaultConfig>;
+export type { DefaultConfig };
 
 /* eslint-disable @typescript-eslint/no-restricted-imports */
 export * from './dev';
